pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cybersentinel"
version = "0.1.0"
description = "End-to-end purple-team cyber-defense lab with multi-agent orchestration"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "CyberSentinel Team", email = "team@cybersentinel.dev"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.6.0",
    "clickhouse-connect>=0.6.0",
    "neo4j>=5.12.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "langgraph>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "sentence-transformers>=2.2.0",
    "protobuf>=4.24.0",
    "grpcio-tools>=1.59.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "redis>=5.0.0",
    "cryptography>=41.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
]
kafka = ["kafka-python>=2.0.2"]
pinecone = ["pinecone-client>=2.2.0"]

[project.scripts]
cybersentinel = "scripts.cli:main"
cs-replay = "scripts.replay:main"
cs-eval = "scripts.eval:main"

[project.urls]
Homepage = "https://github.com/cybersentinel/cybersentinel"
Repository = "https://github.com/cybersentinel/cybersentinel"
Documentation = "https://cybersentinel.github.io/cybersentinel"
Issues = "https://github.com/cybersentinel/cybersentinel/issues"

[tool.setuptools.packages.find]
exclude = ["tests*", "integration_tests*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # ignore unused imports in __init__.py
"tests/**/*" = ["ARG", "B011"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "faiss.*",
    "neo4j.*",
    "clickhouse_connect.*",
    "nats.*",
    "langgraph.*",
    "langchain.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests", "integration_tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = [".", "bus", "storage", "knowledge", "agents", "ingest", "eval"]
omit = [
    "*/tests/*",
    "*/integration_tests/*",
    "setup.py",
    "*/venv/*",
    "*/.env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "integration_tests"]
skips = ["B101", "B601"]  # Skip assert_used and shell injection (for tests)